remise au propre de category, status, role et gender + traitement des erreurs

+++ BLL/CategoryService/ICategoryService.cs
@@ -5,9 +5,9 @@ namespace BLL.CategoryService
     public interface ICategoryService
     {
         Task<IEnumerable<CategoryFull>> GetAll();
-        Task<CategoryFull> GetById(byte id);
-        Task<byte> Add(AddCategory category);
+        Task<CategoryFull> GetById(int id);
+        Task<int> Add(AddCategory category);
         Task<bool> Update(UpdateCategory category);
-        Task<bool> Delete(byte id);
+        Task<bool> Delete(int id);
     }
 }
