+++ DAL/CategoryRepository/ICategoryRepository.cs
@@ -5,9 +5,9 @@ namespace DAL.CategoryRepository
     public interface ICategoryRepository
     {
         Task<IEnumerable<CategoryFull>> GetAll();
-        Task<CategoryFull> GetById(byte id);
-        Task<byte> Add(AddCategory category);
+        Task<CategoryFull> GetById(int id);
+        Task<int> Add(AddCategory category);
         Task<bool> Update(UpdateCategory category);
-        Task<bool> Delete(byte id);
+        Task<bool> Delete(int id);
     }
 }
