+++ Model/Category/CategoryFull.cs
@@ -2,9 +2,9 @@
 {
     public class CategoryFull
     {
-        public byte IdCategory { get; set; }
+        public int IdCategory { get; set; }
         public string NameCategory { get; set; }
-        public byte MinAge { get; set; }
-        public byte MaxAge { get; set; }
+        public int MinAge { get; set; }
+        public int MaxAge { get; set; }
     }
 }
